feat(calendar): enhance calendar functionality with utility functions
- Moved calendar days generation and event type utilities to utils.js
- Updated CalendarPage to utilize the new generateCalendarDays function
- Added getEventTypeIcon and getEventTypeColor functions for better event management

src/app/dashboard/calendar/page.jsx
src/lib/utils.js
